BUG: _ngff_image_scale_factors: Fix infinite loop and tests
To make sure that `while (sizes_array > double_chunks).any():` does not infinite loop, sizes_array and double_chunks must be guaranteed to be in the same order. The `scale_factor == previous` exit condition is an extra sanity safeguard.

The scale factors produced for (1, 125, 1024, 1024) are weird, but I assume this is intended behaviour